apa: 1.when state is parking err, enter apa 2.set if use uss  default from config 3.fix not fill slot_lat_pos_preference when calc origin target pose

+++ src/modules/apa_function/apa_common/apa_slot_manager.cc
@@ -277,7 +277,7 @@ void ApaSlotManager::ParkingLotCruiseProcess() {
       slot.release_info_.release_state[RULE_BASED_RELEASE] =
           SlotReleaseState::NOT_RELEASE;
       ILOG_INFO << "NOT_RELEASE reason: over max released size "
-                << kMaxSlotReleaseCount;
+                << static_cast<int>(kMaxSlotReleaseCount);
       continue;
     }
